vpx_mem,align_addr: use ~ to create mask
removes the need for an intermediate cast to int, which was missing in
the call added in:
69c5ba1 vpx_mem: Refactor code

quiets a visual studio warning:
C4146: unary minus operator applied to unsigned type, result still
unsigned

+++ vpx_mem/vpx_mem.c
@@ -59,7 +59,7 @@ void *vpx_memalign(size_t align, size_t size) {
 
   addr = malloc((size_t)aligned_size);
   if (addr) {
-    x = align_addr((unsigned char *)addr + ADDRESS_STORAGE_SIZE, (int)align);
+    x = align_addr((unsigned char *)addr + ADDRESS_STORAGE_SIZE, align);
     set_actual_malloc_address(x, addr);
   }
   return x;
